feat(banner): slow website banner (#3482)
while we fix the current issues we can warn people
