+++ web/src/components/app.tsx
@@ -46,6 +46,7 @@ import {
 } from './locale-helpers';
 import { Flags } from '../stores/flags';
 import { ReactLocalization, LocalizationProvider } from '@fluent/react';
+import ErrorSlowBanner from './error-slow-banner/error-slow-banner';
 const rtlLocales = require('../../../locales/rtl.json');
 const ListenPage = React.lazy(
   () => import('./pages/contribution/listen/listen')
@@ -245,6 +246,8 @@ let LocalizedPage: any = class extends React.Component<
                     )
                 )}
             </div>
+
+            <ErrorSlowBanner />
 
             <Switch>
               {[
